Convert to ConsistentHash crate.
Although the ConsistentHash can't be modified, this is an example how
bad disks can be filtered out, or a new ConsistentHash can be created in
place.

This has yet another 2x performance improvement over the last iteration
meaning that this code is 4x faster than my original incarnation. On my 2
year old laptop with a single core, it was able to process 20 million
files per second compared to 5 million files per second before.